Move OldReaderUrlReader into a separate file
